requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
chinesecalendar==1.6.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==2.2.7
django-cors-headers==3.10.0
django-mssql==1.8
django-mssql-backend==2.8.1
django-rest-swagger==2.2.0
django-sqlserver==1.11
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
fonttools==4.28.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
openapi-codec==1.3.2
packaging==21.3
pandas==1.0.3
Pillow==8.4.0
PyJWT==1.7.1
pyodbc==4.0.32
pyparsing==3.0.6
python-dateutil==2.8.2
python-tds==1.11.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
setuptools-scm==6.3.2
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
